fcos/requirements.txt
numpy>=1.18.5
Keras==2.2.4
keras-resnet==0.2.0
opencv-contrib-python>=3.4.2.17
opencv-python>=3.4.2.17
Pillow>=6.2.0
tensorflow-gpu==1.13.1
tensorflow==1.13.1
h5py==2.10.0
PyYAML==5.3.1
onnx==1.7.0
onnxruntime
cython
pandas
matplotlib
tqdm
progressbar2
git+https://github.com/cocodataset/cocoapi.git#subdirectory=PythonAPI
